requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.0.0
xgboost>=1.6.0
catboost>=1.0.0
lightgbm>=3.3.0

# Hyperparameter optimization
optuna>=3.0.0

# Feature importance and interpretability
shap>=0.41.0

# Geospatial processing
rasterio>=1.2.0
geopandas>=0.10.0
pyproj>=3.2.0
shapely>=1.8.0

# Climate data access
cdsapi>=0.5.0  # For ERA5 data download
netCDF4>=1.5.8
xarray>=0.20.0
cfgrib>=0.9.10  # For GRIB file reading

# Data visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0
folium>=0.12.0  # For interactive maps

# API access
requests>=2.26.0
aiohttp>=3.8.0  # Async HTTP for PurpleAir API

# Configuration management
pyyaml>=5.4.0
python-dotenv>=0.19.0

# Progress bars and utilities
tqdm>=4.62.0
joblib>=1.1.0

# Notebook support
jupyter>=1.0.0
jupyterlab>=3.0.0
ipywidgets>=7.6.0

# Statistical analysis
statsmodels>=0.13.0

# Quality control
pytest>=6.2.0
pytest-cov>=3.0.0
black>=21.0
flake8>=3.9.0
mypy>=0.910

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0
nbsphinx>=0.8.0  # For notebook documentation

# Time series utilities
pytz>=2021.3

# Logging and monitoring
loguru>=0.5.3

# Web interface (optional, for local deployment)
gradio>=3.0.0
streamlit>=1.10.0

# Parallel processing
dask[complete]>=2021.10.0

# File I/O
h5py>=3.6.0
tables>=3.7.0

# Optional: GPU support for XGBoost/LightGBM
# Uncomment if using GPU acceleration
# cupy-cuda11x>=10.0.0  # Replace 11x with your CUDA version
